requirements.txt
flask
nose
pil
pygame
pynmea2
pyserial
scipy
SimpleCV
